% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateTemporalSmoothness.R
\name{calculateTemporalSmoothness}
\alias{calculateTemporalSmoothness}
\title{Calculate Temporal Smoothness Metric and Return Individual Scores}
\usage{
calculateTemporalSmoothness(cluster_assignments, time_points)
}
\arguments{
\item{cluster_assignments}{A vector of cluster labels.}

\item{time_points}{A vector of time points corresponding to each sample.}
}
\value{
A data frame with the temporal smoothness scores for each pair of time points.
}
\description{
This function computes the Temporal Smoothness Metric (TSM) for each pair of consecutive time points.
}
\examples{
cluster_assignments <- c(1, 1, 2, 2, 3, 3, 1, 1, 2, 2)
time_points <- c(1, 1, 2, 2, 3, 3, 4, 4, 5, 5)
tsm_df <- calculateTemporalSmoothness(cluster_assignments, time_points)
}
